# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: PrestaRapi - Sistema de Gestión de Préstamos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List users
      parameters:
        - name: agencyId
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Soft-delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/{id}/toggle-block:
    post:
      operationId: toggleUserBlock
      tags: [users]
      summary: Block or unblock a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User block status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /payments:
    get:
      operationId: getPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: agencyId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: loanId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: registerPayment
      tags: [payments]
      summary: Register a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPaymentRequest"
      responses:
        "201":
          description: Payment registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}/reconcile:
    post:
      operationId: reconcilePayment
      tags: [payments]
      summary: Reconcile a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconcileRequest"
      responses:
        "200":
          description: Payment reconciled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}/reverse:
    post:
      operationId: reversePayment
      tags: [payments]
      summary: Reverse a reconciled payment (SuperAdmin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment reversed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
  /movements:
    get:
      operationId: getMovements
      tags: [payments]
      summary: List loan movements
      parameters:
        - name: loanId
          in: query
          schema:
            type: integer
        - name: agencyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movement"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [superadmin, admin, manager, promoter]
        agencyIds:
          type: array
          items:
            type: integer
        blocked:
          type: boolean
        avatar:
          type: string
        needsPasswordRecovery:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - username
        - role
        - agencyIds
        - blocked
        - needsPasswordRecovery
        - createdAt
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, promoter]
        agencyIds:
          type: array
          items:
            type: integer
      required:
        - name
        - username
        - password
        - role
        - agencyIds
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        agencyIds:
          type: array
          items:
            type: integer
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        loanId:
          type: integer
        clientName:
          type: string
        agencyId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending_reconciliation, reconciled, reversed]
        registeredById:
          type: integer
        registeredByName:
          type: string
        reconciledById:
          type: integer
        reconciledByName:
          type: string
        image:
          type: string
        date:
          type: string
          format: date
        appliedOverdue:
          type: number
        appliedCurrent:
          type: number
        appliedCapital:
          type: number
        forgivenInterest:
          type: number
        forgivenCapital:
          type: number
      required:
        - id
        - loanId
        - agencyId
        - amount
        - status
        - registeredById
        - date
    RegisterPaymentRequest:
      type: object
      properties:
        loanId:
          type: integer
        amount:
          type: number
        image:
          type: string
          description: Base64 encoded image or URL
        notes:
          type: string
      required:
        - loanId
        - amount
    ReconcileRequest:
      type: object
      properties:
        forgivenInterest:
          type: number
          description: Amount of interest to forgive
        forgivenCapital:
          type: number
          description: Amount of capital to forgive
      required: []
...
    Movement:
      type: object
      properties:
        id:
          type: integer
        loanId:
          type: integer
        type:
          type: string
          enum: [disbursement, interest_accrual, payment_overdue_interest, payment_current_interest, payment_capital, forgiveness, reversal]
        amount:
          type: number
        concept:
          type: string
        date:
          type: string
          format: date
        userId:
          type: integer
        userName:
          type: string
        reversed:
          type: boolean
        image:
          type: string
      required:
        - id
        - loanId
        - type
        - amount
        - concept
        - date
        - reversed